moved windows placing
