+++ src/modules/Workspaces/WorkspacesWindowArranger/ArrangeWindows.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <vector>
+
+#include <WorkspacesLib/WorkspacesData.h>
+
+bool ArrangeWindows(WorkspacesData::WorkspacesProject& project, const std::vector<WorkspacesData::WorkspacesProject::Monitor>& monitors);
