+++ src/modules/Workspaces/WorkspacesWindowArranger/LaunchingApp.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <Windows.h>
+#include <WorkspacesLib/WorkspacesData.h>
+
+struct LaunchingApp
+{
+    WorkspacesData::WorkspacesProject::Application application;
+    HWND window;
+    std::wstring state;
+};
+
+using LaunchingApps = std::vector<LaunchingApp>;
