+++ src/modules/Workspaces/WorkspacesWindowArranger/WorkspacesWindowArranger.vcxproj
@@ -126,20 +126,17 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="UIHelper.cpp" />
+    <ClCompile Include="ArrangeWindows.cpp" />
     <ClCompile Include="main.cpp" />
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader Condition="'$(UsePrecompiledHeaders)' != 'false'">Create</PrecompiledHeader>
     </ClCompile>
-    <ClCompile Include="RegistryUtils.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="UIHelper.h" />
+    <ClInclude Include="ArrangeWindows.h" />
     <ClInclude Include="LaunchingApp.h" />
     <ClInclude Include="pch.h" />
-    <ClInclude Include="RegistryUtils.h" />
     <ClInclude Include="resource.h" />
-    <ClInclude Include="utils.h" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
